share app client hydration and init code
