+++ packages/next-swc/crates/next-core/src/next_server/context.rs
@@ -240,6 +240,7 @@ fn defines(mode: NextMode) -> CompileTimeDefines {
         process.turbopack = true,
         process.env.NODE_ENV = mode.node_env(),
         process.env.__NEXT_CLIENT_ROUTER_FILTER_ENABLED = false,
+        process.env.__NEXT_NEW_LINK_BEHAVIOR = true,
         process.env.NEXT_RUNTIME = "nodejs"
     )
     // TODO(WEB-937) there are more defines needed, see
