+++ packages/next/src/client/app-next-dev.ts
@@ -3,6 +3,8 @@
 import { appBootstrap } from './app-bootstrap'
 
 appBootstrap(() => {
+  const { install } = require('./app-webpack')
+  install()
   const { hydrate } = require('./app-index')
   hydrate()
 })
