+++ packages/next/src/client/app-next.ts
@@ -4,6 +4,7 @@ appBootstrap(() => {
   // Include app-router and layout-router in the main chunk
   require('next/dist/client/components/app-router')
   require('next/dist/client/components/layout-router')
+  require('./app-webpack')
   const { hydrate } = require('./app-index')
   hydrate()
 })
